+++ prober.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "1.14.e"
+__version__ = "1.15.a"
 # Year, month, day
-__last_update_date__ = "2020-01-13"
+__last_update_date__ = "2020-01-28"
 
 # |===== Check python interpreter version =====|
 
@@ -333,6 +333,8 @@ import shelve
 import multiprocessing as mp
 import threading
 
+from math import log
+
 # |===== Functionality for proper processing of gzipped files =====|
 
 OPEN_FUNCS = (open, open_as_gzip)
@@ -584,13 +586,21 @@ Enter a number (1 or 2):>> """)
 
 # |===== End of question funtions =====|
 
-get_phred33 = lambda q_symb: ord(q_symb) - 33
+# Function for getting Q value from Phred33 character:
+substr_phred33 = lambda q_symb: ord(q_symb) - 33
+# List of propabilities corresponding to indices (index is Q, is the propability):
+q2p_map = [10 ** (-q/10) for q in range(128)] # 127 -- max value of a signed byte
+# Function for accessing propabilities by Q:
+qual2prop = lambda q: q2p_map[q]
+# Function for accessing Q by propability:
+prop2qual = lambda p: round(-10 * log(p, 10), 2)
 
 def get_read_avg_qual(qual_str):
 
-    phred33 = map(get_phred33, list(qual_str))
-    read_qual = round( sum(phred33) / len(qual_str), 2 )
-    return read_qual
+    quals = map(substr_phred33, qual_str) # get Qs
+    err_props = map(qual2prop, quals) # convert Qs to propabilities
+    avg_err_prop = sum(err_props) / len(qual_str) # calculate average propability
+    return prop2qual(avg_err_prop)
 # end def get_read_avg_qual
 
 
@@ -602,6 +612,8 @@ def configure_qual_dict(fastq_path):
     :param fastq_path: path to FASTQ file meant to be processed;
     :type fastq_path: str;
     """
+
+    printl("\n{} - Calculating mean read qualities...".format(get_work_time()))
 
     qual_dict = dict()
     how_to_open = OPEN_FUNCS[ is_gzipped(fastq_path) ]
@@ -623,6 +635,8 @@ def configure_qual_dict(fastq_path):
             # end if
         # end while
     # end with
+
+    printl("{} - Done.".format(get_work_time()))
 
     return qual_dict
 # end def configure_qual_dict
@@ -844,7 +858,7 @@ def look_around(outdir_path, new_dpath, fasta_path, blast_algorithm):
             except Exception as err:
                 printl("\nData in accession file '{}' not found or broken. Reason:".format(acc_fpath))
                 printl( ' ' + str(err) )
-                printl("Start from the beginning.")
+                printl("Starting from the beginning.")
                 rename_file_verbosely(tsv_res_fpath, new_dpath)
                 rename_file_verbosely(tmp_fpath, new_dpath)
                 rename_file_verbosely(acc_fpath, new_dpath)
@@ -1001,7 +1015,7 @@ def fasta_packets(fasta, packet_size, reads_at_all, num_done_reads):
         packet += line+'\n' # add recently read line
 
         packs_at_all = reads_at_all // packet_size # Calculate total number of packets sent from current FASTA file
-        if reads_at_all % packet_size > 0: # And this is ceiling (in order not to import 'math')
+        if reads_at_all % packet_size > 0: # And this is ceiling
             packs_at_all += 1
         # end if
         packs_processed = int( num_done_reads / packet_size ) # number of successfully processed sequences
@@ -1489,23 +1503,47 @@ def get_lineage(gi, hit_def, hit_acc):
             spec_name = re_search(r" ([a-z]+(\. [a-zA-Z0-9]+)?)", org_name).group(1)
 
             # Get full lineage
-            lin_regex = r"<INSDSeq_taxonomy>([A-Za-z; ]+)</INSDSeq_taxonomy>"
+            lin_regex = r"<INSDSeq_taxonomy>([A-Za-z0-9;\. ]+)</INSDSeq_taxonomy>"
             lineage = re_search(lin_regex, text).group(1).strip('.')
 
-            # Check if species name is in lineage:
-            gen_spec_regex = r"([A-Z][a-z]+ [a-z]+).?$"
-            gen_spec_in_lin = re_search(gen_spec_regex, lineage)
+            # Format of genus-species in lineage can vary.
+            # We need to parse it correctly anyway.
+            if spec_name != "sp": # no species info will be added to lineage for "Pseudarthrobacter sp."
 
-            # All taxonomy names in lineage will be divided by semicolon:
-            if not gen_spec_in_lin is None:
-                repl_str = gen_spec_in_lin.group(1)
-                lineage = lineage.replace(repl_str, spec_name)
-            else:
-                lineage += ";" + spec_name
-            # emd if
+                # Remove "Bacillus cereus group" from lineage
+                for grp_cmplx in (" group", " complex"):
+                    if grp_cmplx in lineage:
+                        lineage = lineage[: lineage.rfind(';')]
+                    # end if
+                # end for
+
+                # Check if genus and species names are in lineage (separated by space):
+                gen_spec_regex = r"(([A-Z][a-z]+) [a-z]+(\. [a-zA-Z0-9]+)?).?$"
+                gen_spec_in_lin = re_search(gen_spec_regex, lineage)
+
+                # If there is genus and species in lineage
+                if not gen_spec_in_lin is None:
+                    genus_name = gen_spec_in_lin.group(2)
+                    genus_count = lineage.count(" " + genus_name) # "...;Bacillus; Bacillus subtilis" are not allowed
+                    if genus_count == 1:
+                        pass # leave genus and species names intact
+                    elif genus_count == 2: # remove odd genus name
+                        lineage = lineage.replace(" " + genus_name + " ", " ")
+                    else:
+                        print(err_fmt("taxonomy parsing error"))
+                        print("Please, contact the developer.")
+                        print("Tell him that 'genus_count' is {}".format(genus_count))
+                        platf_depend_exit(1)
+                    # end if
+                elif not ' ' + spec_name in lineage:
+                    lineage += ";" + spec_name # if there are no species name -- add it
+                # end if
+            # end if
 
             # Remove all spaces:
-            lineage = lineage.replace(' ', '')
+            lineage = lineage.replace("sp. ", "sp._")
+            lineage = lineage.replace("; ", ";")
+            lineage = lineage.replace(" ", ";") # in orger to separate genus and species with semicolon
 
         except AttributeError:
             # If there is no lineage -- use hit definition instead of it
@@ -1947,7 +1985,7 @@ for i, fq_fa_path in enumerate(fq_fa_list):
     tmp_num = min(probing_batch_size, curr_fasta["nreads"])
     packs_at_all = tmp_num // packet_size
 
-    if tmp_num % packet_size > 0: # And this is ceiling (in order not to import 'math')
+    if tmp_num % packet_size > 0: # And this is ceiling
         packs_at_all += 1
     # end if
 
@@ -2076,7 +2114,7 @@ for acc, other_info in sorted(acc_dict.items(), key=lambda x: -x[1][2]):
 unkn_num = glob_seqs_processed - sum( map(lambda x: x[2], acc_dict.values()) )
 if unkn_num > 0:
     s_letter = "s" if unkn_num > 1 else ""
-    printl(" {} hit{} - No significant similarity found".format(unkn_num, s_letter))
+    printl(" {} sequenc{} - No significant similarity found".format(unkn_num, s_letter))
 # end if
 
 printl("""\nThey are saved in following file: